Rename ert/data/measured->ert/data/_measured_data
